requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.4
Flask-Mail==0.10.0
SQLAlchemy==2.0.35
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
python-dotenv==1.0.1
gunicorn==22.0.0
eventlet==0.33.3
qrcode==7.4.2
Pillow==10.4.0
Flask-JWT-Extended==4.5.3
Werkzeug==3.0.3
itsdangerous==2.2.0
psycopg2-binary==2.9.10
